fix: removed unnecessary code; moved consts to shared/const
