+++ Weather/index.html
@@ -12,8 +12,8 @@
     />
     <title>Weather</title>
   </head>
-  <body class="flex flex-col" id="body" data-theme="light">
-    <div id="root" class="flex flex-col w-full h-full"></div>
+  <body id="body" data-theme="light">
+    <div id="root"></div>
     <script type="module" src="/src/main.jsx"></script>
   </body>
 </html>
