+++ Weather/src/App.scss
@@ -46,6 +46,7 @@ body::before {
     rgba(255, 255, 255, 0.4) -81.35%,
     rgba(255, 255, 255, 0) 112.36%
   );
+  --search-icon: url("./public/icons/search_dark.svg");
   --background-color: linear-gradient(
     162.91deg,
     rgba(255, 255, 255, 0.03) -1.68%,
@@ -57,6 +58,9 @@ body::before {
 
   background-color: #080338;
   color: var(--main-color);
+  input{
+    color: var(--main-color);
+  }
 }
 
 [data-theme="light"] {
@@ -66,6 +70,7 @@ body::before {
     rgba(73, 73, 73, 0.4) -81.35%,
     rgba(255, 255, 255, 0) 112.36%
   );
+  --search-icon: url("./public/icons/search_light.svg");
   --background-color: linear-gradient(
     162.91deg,
     rgba(0, 0, 0, 0.03) -1.68%,
@@ -77,6 +82,9 @@ body::before {
 
   background-color: #fff;
   color: var(--main-color);
+  input{
+    color: var(--main-color);
+  }
 }
 
 #root {
