+++ Weather/src/components/Days/Days.scss
@@ -6,7 +6,7 @@
     justify-content: space-between;
     gap: 2rem;
     .day__item{
-        height: 150px;
+        min-height: 9rem;
         width: 100%;
     }
 }
