+++ Weather/src/components/Header/Header.scss
@@ -1,6 +1,6 @@
 .header {
   width: 100%;
-  .header__content{
+  .header__content {
     display: flex;
     align-items: center;
     justify-content: space-between;
@@ -10,12 +10,12 @@
     font-family: Open Sans;
     font-weight: 400;
     font-style: Regular;
-    font-size: 24px;
+    font-size: 1.5rem;
     line-height: 100%;
     letter-spacing: 0%;
   }
 
-  .header__functions{
+  .header__functions {
     display: flex;
     align-items: center;
     justify-content: space-between;
@@ -29,8 +29,8 @@
     gap: 1rem;
     position: relative;
     outline: none;
-    width: 360px;
-    height: 50px;
+    width: 22.5rem;
+    height: 3rem;
     border-radius: 30px;
     padding: 0 20px;
     background: var(--background-color);
@@ -49,21 +49,35 @@
       mask-composite: exclude;
       pointer-events: none;
     }
-    .header__search-icon path {
-      fill: var(--main-color);
+
+    .header__search-icon {
+      background: var(--search-icon) no-repeat center center / contain;
+      width: 1.5rem;
+      height: 1.5rem;
+      display: inline-block;
     }
+  }
+
+  .header__container .input-field {
+    width: 100%;
+    height: 100%;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    background: transparent;
+    border: none;
   }
 }
 
 .header__container.header__button {
-  width: 75px;
+  width: 5rem;
   overflow: hidden;
   z-index: 1;
   &.light {
     &::after {
-      left: 25px;
-      background: #fff url("./public/icons/cloud+sun.png") no-repeat center 
-      center / 70% auto;;
+      left: 2rem;
+      background: #fff url("./public/icons/cloud+sun.png") no-repeat center
+        center / 70% auto;
     }
   }
   &::after {
@@ -73,10 +87,10 @@
     position: absolute;
     content: "";
     inset: 0;
-    width: 50px;
+    width: 3rem;
     height: 100%;
     border-radius: 100%;
-    background: #080338 url("./public/icons/cloud+moon.png") no-repeat center 
+    background: #080338 url("./public/icons/cloud+moon.png") no-repeat center
       center / 70% auto;
   }
 }
