+++ Weather/src/components/Today/Today.scss
@@ -1,5 +1,5 @@
 .today{
     div{
-        height: 250px;
+        min-height: 18rem;
     }
 }
