+++ Weather/src/index.css
@@ -1,5 +1,3 @@
-@import "tailwindcss";
-
 :root {
   font-family: "Open Sans", sans-serif;
   line-height: 1.5;
@@ -21,6 +19,15 @@ body {
   place-items: center;
   min-width: 20rem;
   min-height: 100vh;
+  display: flex;
+  flex-direction: column;
+}
+
+#root{
+  display: flex;
+  flex-direction: column;
+  width: 100%;
+  height: 100%;
 }
 
 h1 {
