+++ Weather/src/shared/const/DaysData.jsx
@@ -1,0 +1,29 @@
+const DaysData = [
+    {
+        title: "Tuesday",
+
+    },
+    {
+        title: "Wednesday",
+
+    },
+    {
+        title: "Thursday",
+
+    },
+    {
+        title: "Friday",
+
+    },
+    {
+        title: "Saturday",
+
+    },
+    {
+        title: "Sunday",
+
+    }
+]
+
+export default DaysData;
+
