+++ Weather/src/shared/const/GraphsData.jsx
@@ -1,0 +1,13 @@
+const graphsData = [
+  {
+    title: "UV Index",
+  },
+  {
+    title: "Sunrise  and Sunset",
+  },
+  {
+    title: "Humidity",
+  },
+];
+
+export default graphsData;
